Modify the TypeTester interface
The TypeTester interface has been modified so that it returns an optional string instead of a boolean. For a single test, boolean is fine, but once the code is generalized, we won't know the type that is being tested for (and conceivably, one test might test for multiple types), so it is not sufficient to say the test passed.

We must instead indicate that the test passes *and* what type it passes for.

Tests have been updated accordingly.

Eventually, we will probably need to change the type indicator from a String to something else more type safe, but it's fine for now.