Refactor tests using a try catch block and parser
We will be testing integers and some other primitives by simply calling built-in java parse code. There is no point in writing the try/catch code for all of these types, though. Instead, I have introduced a TryCatchTester which takes a function provided by the user and applies it to the sample within a try/catch block. To create new tests of this format, we just need to provide a type string and a parsing function.