Add Test Adapters
Adding test adapters to allow for the use of tests in different contexts. At the moment I have used them to duplicate the aggregate and stream versions of the test code in the TypeTester interface (I'll probably remove that functionality from that interface).

The motivation for this is basically that if we want to drop this into, for example, Apache Spark, we would like to work with their RDD type instead of Java streams or lists. So, we would create a simple wrapper that handles using RDDs. For other frameworks, we might do something similar. Since we'll need to do this, it seems to make sense to just formalize these wrappers and keep the TypeTester simple.