Add DocumentAdapter
Added an abstract DocumentAdapter class which allows for the creation of adapters for different types of documents. The user must simply implement a method that converts the document to a stream of samples, then we delegate to a stream adapter to execute the tests on all values in the document.